refactor(Message): 에러 메시지 통합 및 관리 일원화
- 에러 메시지를 Message enum으로 통합하여 일관된 메시지 관리 구현
- 숫자 처리 메서드를 appendNumber로 변경하여 명확성 향상
